refactor: move module tests to their own subdirectory
